refactor: move db to common
