+++ src/users/users.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { UsersRepository } from '../DB/users-db';
+import { UsersRepository } from '../common/DB/users-db';
 
 @Injectable()
 export class UsersService {
